+++ R/canonical.R
@@ -8,20 +8,7 @@
 #' @param .dag,.tdy_dag input graph, an object of class `tidy_dagitty` or
 #'   `dagitty`
 #' @param ... additional arguments passed to `tidy_dagitty()`
-#' @param edge_type a character vector, the edge geom to use. One of:
-#'   "link_arc", which accounts for directed and bidirected edges, "link",
-#'   "arc", or "diagonal"
-#' @param node_size size of DAG node
-#' @param text_size size of DAG text
-#' @param label_size size of label text
-#' @param text_col color of DAG text
-#' @param label_col color of label text
-#' @param node logical. Should nodes be included in the DAG?
-#' @param stylized logical. Should DAG nodes be stylized? If so, use
-#'   `geom_dag_nodes` and if not use `geom_dag_point`
-#' @param text logical. Should text be included in the DAG?
-#' @param use_labels a string. Variable to use for `geom_dag_label_repel()`.
-#'   Default is `NULL`.
+#' @inheritParams geom_dag
 #'
 #' @return a `tidy_dagitty` that includes L or a `ggplot`
 #' @export
@@ -45,14 +32,16 @@ node_canonical <- function(.dag, ...) {
 #' @export
 ggdag_canonical <- function(.tdy_dag, ..., edge_type = "link_arc", node_size = 16, text_size = 3.88,
                             label_size = text_size,
-                            text_col = "white", label_col = text_col,
-                            node = TRUE, stylized = FALSE, text = TRUE,
-                            use_labels = NULL) {
+                            text_col = "white", label_col = text_col, use_edges = TRUE,
+                            use_nodes = TRUE, use_stylized = FALSE, use_text = TRUE,
+                            use_labels = NULL, label = NULL, text = NULL, node = deprecated(),
+                            stylized = deprecated()) {
   if_not_tidy_daggity(.tdy_dag, ...) %>%
     node_canonical() %>%
     ggdag(
       node_size = node_size, text_size = text_size, label_size,
       edge_type = edge_type, text_col = text_col, label_col = label_col,
-      node = node, stylized = stylized, text = text, use_labels = use_labels
+      use_edges = use_edges, use_nodes = use_nodes, use_stylized = use_stylized,
+      use_text = use_text, use_labels = use_labels
     )
 }