+++ R/drelationship.R
@@ -11,26 +11,13 @@
 #' @param .tdy_dag input graph, an object of class `tidy_dagitty` or
 #'   `dagitty`
 #' @param ... additional arguments passed to `tidy_dagitty()`
-#' @param edge_type a character vector, the edge geom to use. One of:
-#'   "link_arc", which accounts for directed and bidirected edges, "link",
-#'   "arc", or "diagonal"
 #' @param from a character vector, the starting variable (must by in DAG). If
 #'   `NULL`, checks DAG for exposure variable.
 #' @param to a character vector, the ending variable (must by in DAG). If
 #'   `NULL`, checks DAG for outcome variable.
 #' @param controlling_for a character vector, variables in the DAG to control
 #'   for.
-#' @param node_size size of DAG node
-#' @param text_size size of DAG text
-#' @param label_size size of label text
-#' @param text_col color of DAG text
-#' @param label_col color of label text
-#' @param node logical. Should nodes be included in the DAG?
-#' @param stylized logical. Should DAG nodes be stylized? If so, use
-#'   `geom_dag_nodes` and if not use `geom_dag_point`
-#' @param text logical. Should text be included in the DAG?
-#' @param use_labels a string. Variable to use for `geom_dag_label_repel()`.
-#'   Default is `NULL`.
+#' @inheritParams geom_dag
 #' @param collider_lines logical. Should the plot show paths activated by
 #'   adjusting for a collider?
 #' @param as_factor logical. Should the `d_relationship` variable be a
@@ -194,89 +181,110 @@ node_drelationship <- function(.tdy_dag, from = NULL, to = NULL, controlling_for
       TRUE ~ NA_character_
     )
   )
-if (as_factor) {
-  .tdy_dag <- dplyr::mutate(
-    .tdy_dag,
-    d_relationship = factor(
-      d_relationship,
-      levels = c("d-connected", "d-separated"),
-      exclude = NA
+  if (as_factor) {
+    .tdy_dag <- dplyr::mutate(
+      .tdy_dag,
+      d_relationship = factor(
+        d_relationship,
+        levels = c("d-connected", "d-separated"),
+        exclude = NA
+      )
     )
-  )
-}
-.tdy_dag
+  }
+  .tdy_dag
 }
 
 #' @rdname d_relationship
 #' @export
-ggdag_drelationship <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ..., edge_type = "link_arc",
-                                node_size = 16, text_size = 3.88, label_size = text_size,
-                                text_col = "white", label_col = text_col,
-                                node = TRUE, stylized = FALSE, text = TRUE, use_labels = NULL, collider_lines = TRUE) {
-  edge_function <- edge_type_switch(edge_type)
+ggdag_drelationship <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ...,
+                                edge_type = "link_arc", size = 1,
+                                node_size = 16, text_size = 3.88,
+                                label_size = text_size,
+                                text_col = "white", label_col = "black",
+                                edge_width = 0.6, edge_cap = 10, arrow_length = 5,
+                                use_edges = TRUE,
+                                use_nodes = TRUE, use_stylized = FALSE, use_text = TRUE,
+                                use_labels = FALSE, label = NULL, text = NULL, node = deprecated(),
+                                stylized = deprecated(), collider_lines = TRUE) {
 
   p <- if_not_tidy_daggity(.tdy_dag) %>%
     node_drelationship(from = from, to = to, controlling_for = controlling_for, ...) %>%
-    ggplot2::ggplot(ggplot2::aes(x = x, y = y, xend = xend, yend = yend, shape = adjusted, col = d_relationship)) +
-    edge_function(
-      start_cap = ggraph::circle(10, "mm"),
-      end_cap = ggraph::circle(10, "mm")
+    ggplot2::ggplot(aes_dag(shape = adjusted, col = d_relationship))
+
+    if (collider_lines) p <- p + geom_dag_collider_edges()
+
+    p <- p + geom_dag(
+      size = size,
+      edge_type = edge_type,
+      node_size = node_size,
+      text_size = text_size,
+      label_size = label_size,
+      text_col = text_col,
+      label_col = label_col,
+      edge_width = edge_width,
+      edge_cap = edge_cap,
+      arrow_length = arrow_length,
+      use_edges = use_edges,
+      use_nodes = use_nodes,
+      use_stylized = use_stylized,
+      use_text = use_text,
+      use_labels = use_labels,
+      text = !!rlang::enquo(text),
+      label = !!rlang::enquo(label),
+      node = node,
+      stylized = stylized
     ) +
     scale_adjusted() +
     breaks(c("d-connected", "d-separated"), name = "d-relationship") +
     expand_plot(expand_y = expansion(c(0.2, 0.2)))
 
-  if (collider_lines) p <- p + geom_dag_collider_edges()
-  if (node) {
-    if (stylized) {
-      p <- p + geom_dag_node(size = node_size)
-    } else {
-      p <- p + geom_dag_point(size = node_size)
-    }
-  }
-
-  if (text) p <- p + geom_dag_text(col = text_col, size = text_size)
-
-  if (!is.null(use_labels)) {
-    p <- p +
-      geom_dag_label_repel(
-        ggplot2::aes(
-          label = !!rlang::sym(use_labels),
-          fill = d_relationship
-        ),
-        size = text_size,
-        col = label_col, show.legend = FALSE
-      )
-  }
   p
 }
 
 #' @rdname d_relationship
 #' @export
-ggdag_dseparated <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ..., edge_type = "link_arc",
-                             node_size = 16, text_size = 3.88, label_size = text_size,
-                             text_col = "white", label_col = text_col,
-                             node = TRUE, stylized = FALSE, text = TRUE, use_labels = NULL, collider_lines = TRUE) {
+ggdag_dseparated <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ...,
+                             edge_type = "link_arc", size = 1,
+                             node_size = 16, text_size = 3.88,
+                             label_size = text_size,
+                             text_col = "white", label_col = "black",
+                             edge_width = 0.6, edge_cap = 10, arrow_length = 5,
+                             use_nodes = TRUE, use_stylized = FALSE, use_text = TRUE,
+                             use_labels = FALSE, label = NULL, text = NULL, node = deprecated(),
+                             stylized = deprecated(), collider_lines = TRUE) {
   ggdag_drelationship(
-    .tdy_dag = .tdy_dag, from = from, to = to,
-    controlling_for = controlling_for, ..., edge_type = edge_type,
-    node_size = node_size, text_size = text_size, label_size = label_size,
-    text_col = text_col, label_col = label_col, node = node, stylized = stylized, text = text,
-    use_labels = use_labels, collider_lines = collider_lines
+    .tdy_dag = .tdy_dag, from = from, to = to, controlling_for = controlling_for, ...,
+    edge_type = edge_type, size = size,
+    node_size = node_size, text_size = text_size,
+    label_size = label_size,
+    text_col = text_col, label_col = label_col,
+    edge_width = edge_width, edge_cap = edge_cap, arrow_length = arrow_length,
+    use_nodes = use_nodes, use_stylized = use_stylized, use_text = use_text,
+    use_labels = use_labels, label = !!rlang::enquo(label), text = !!rlang::enquo(text), node = node,
+    stylized = stylized, collider_lines = collider_lines
   )
 }
 
 #' @rdname d_relationship
 #' @export
-ggdag_dconnected <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ..., edge_type = "link_arc",
-                             node_size = 16, text_size = 3.88, label_size = text_size,
-                             text_col = "white", label_col = text_col,
-                             node = TRUE, stylized = FALSE, text = TRUE, use_labels = NULL, collider_lines = TRUE) {
+ggdag_dconnected <- function(.tdy_dag, from = NULL, to = NULL, controlling_for = NULL, ...,
+                             edge_type = "link_arc", size = 1,
+                             node_size = 16, text_size = 3.88,
+                             label_size = text_size,
+                             text_col = "white", label_col = "black",
+                             edge_width = 0.6, edge_cap = 10, arrow_length = 5,
+                             use_nodes = TRUE, use_stylized = FALSE, use_text = TRUE,
+                             use_labels = FALSE, label = NULL, text = NULL, node = deprecated(),
+                             stylized = deprecated(), collider_lines = TRUE) {
   ggdag_drelationship(
-    .tdy_dag = .tdy_dag, from = from, to = to,
-    controlling_for = controlling_for, ..., edge_type = edge_type,
-    node_size = node_size, text_size = text_size, label_size = label_size,
-    text_col = text_col, label_col = label_col, node = node, stylized = stylized, text = text,
-    use_labels = use_labels, collider_lines = collider_lines
+    .tdy_dag = .tdy_dag, from = from, to = to, controlling_for = controlling_for, ...,
+    edge_type = edge_type, size = size,
+    node_size = node_size, text_size = text_size,
+    label_size = label_size,
+    text_col = text_col, label_col = label_col,
+    edge_width = edge_width, edge_cap = edge_cap, arrow_length = arrow_length,
+    use_nodes = use_nodes, use_stylized = use_stylized, use_text = use_text,
+    use_labels = use_labels, label = !!rlang::enquo(label), text = !!rlang::enquo(text), node = node,
+    stylized = stylized, collider_lines = collider_lines
   )
 }