+++ R/ggdag.R
@@ -5,20 +5,7 @@
 #' @param .tdy_dag input graph, an object of class `tidy_dagitty` or
 #'   `dagitty`
 #' @param ... additional arguments passed to `tidy_dagitty()`
-#' @param edge_type a character vector, the edge geom to use. One of:
-#'   "link_arc", which accounts for directed and bidirected edges, "link",
-#'   "arc", or "diagonal"
-#' @param node_size size of DAG node
-#' @param text_size size of DAG text
-#' @param label_size size of label text
-#' @param text_col color of DAG text
-#' @param label_col color of label text
-#' @param node logical. Should nodes be included in the DAG?
-#' @param stylized logical. Should DAG nodes be stylized? If so, use
-#'   `geom_dag_nodes` and if not use `geom_dag_point`
-#' @param text logical. Should text be included in the DAG?
-#' @param use_labels a string. Variable to use for `geom_dag_label_repel()`.
-#'   Default is `NULL`.
+#' @inheritParams geom_dag
 #'
 #' @return a `ggplot`
 #' @export
@@ -30,46 +17,48 @@
 #'   x ~ z1 + w1,
 #'   z1 ~ w1 + v,
 #'   z2 ~ w2 + v,
-#'   w1 ~ ~w2
+#'   w1 ~~ w2
 #' )
 #'
 #' ggdag(dag)
-#' ggdag(dag) + theme_dag_blank()
+#' ggdag(dag) + theme_dag()
 #'
 #' ggdag(dagitty::randomDAG(5, .5))
 #'
 #' @seealso [ggdag_classic()]
-ggdag <- function(.tdy_dag, ..., edge_type = "link_arc", node_size = 16, text_size = 3.88,
-                  label_size = text_size,
+ggdag <- function(.tdy_dag, ...,
+                  size = 1, edge_type = c("link_arc", "link", "arc", "diagonal"),
+                  node_size = 16, text_size = 3.88, label_size = text_size,
                   text_col = "white", label_col = "black",
-                  node = TRUE, stylized = FALSE, text = TRUE,
-                  use_labels = NULL) {
-  edge_function <- edge_type_switch(edge_type)
+                  edge_width = 0.6, edge_cap = 8, arrow_length = 5,
+                  use_edges = TRUE,
+                  use_nodes = TRUE, use_stylized = FALSE, use_text = TRUE,
+                  use_labels = FALSE, text = NULL, label = NULL,
+                  node = deprecated(), stylized = deprecated()) {
 
-  p <- if_not_tidy_daggity(.tdy_dag, ...) %>%
-    ggplot2::ggplot(ggplot2::aes(x = x, y = y, xend = xend, yend = yend)) +
-    edge_function()
-
-  if (node) {
-    if (stylized) {
-      p <- p + geom_dag_node(size = node_size)
-    } else {
-      p <- p + geom_dag_point(size = node_size)
-    }
-  }
-
-  if (text) p <- p + geom_dag_text(col = text_col, size = text_size)
-
-  if (!is.null(use_labels)) {
-    p <- p +
-      geom_dag_label_repel(
-        ggplot2::aes(label = !!rlang::sym(use_labels)),
-        size = text_size,
-        col = label_col,
-        show.legend = FALSE
-      )
-  }
-  p
+  if_not_tidy_daggity(.tdy_dag, ...) %>%
+    ggplot2::ggplot(aes_dag()) +
+    geom_dag(
+      size = size,
+      edge_type = edge_type,
+      node_size = node_size,
+      text_size = text_size,
+      label_size = label_size,
+      text_col = text_col,
+      label_col = label_col,
+      edge_width = edge_width,
+      edge_cap = edge_cap,
+      arrow_length = arrow_length,
+      use_edges = use_edges,
+      use_nodes = use_nodes,
+      use_stylized = use_stylized,
+      use_text = use_text,
+      use_labels = use_labels,
+      text = !!rlang::enquo(text),
+      label = !!rlang::enquo(label),
+      node = node,
+      stylized = stylized
+    )
 }
 
 #' Quickly plot a DAG in ggplot2
@@ -113,7 +102,7 @@ ggdag_classic <- function(.tdy_dag, ..., size = 8, label_rect_size = NULL,
   fontsize <- ifelse(!is.null(label_rect_size), label_rect_size, size * 3.57)
 
   p <- .tdy_dag %>%
-    ggplot2::ggplot(ggplot2::aes(x = x, y = y, xend = xend, yend = yend)) +
+    ggplot2::ggplot(aes_dag()) +
     ggplot2::geom_text(
       ggplot2::aes(label = !!rlang::sym(text_label)),
       size = size,