+++ tests/testthat/test-paths.R
@@ -41,6 +41,10 @@ test_that("DAG paths are identified and drawn correctly", {
   p2 <- butterfly_bias_dag %>%
     ggdag_paths_fan(shadow = TRUE)
 
+  p3 <- confounder_triangle_dag %>%
+    ggdag_paths(from = "x", to = "y", shadow = TRUE)
+
   expect_doppelganger("ggdag_paths() draws 2 open paths", p1)
   expect_doppelganger("ggdag_paths_fan() draws 4 open paths", p2)
+  expect_doppelganger("ggdag_paths() draws 2 open paths with shadows", p3)
 })